Feat: Add random number generation to function
The computer function needs to randomly provide
a value back to be able to use that as a choice.
the number returned currently is between 0 & 1.
I wanted the number to be less accurate and
shorter, so I used the float function on the
value and then had it rounded to two decimal
places using toFixed(2).