Feat: Add a function for the user's choice
Added a function that prompts the user to enter
a choice through a text input. It checks if the
user inputs one of the three correct choices.
This isn't at a stage where it is a catch all for
what can be entered by the user, but it works for
the choices that the game is meant to take and
use.