Feat: Add a play round function
Added a function for playing a round. This is
just a skeleton at this stage while following
theOdinProjects explainations. Importantly logged
the values that the two functions I created to
check whether the new variables that are assigned
them work. If the user inputs a non
answer, it returns undefined. I can utilise this
for the game as an error situation.