Feat: Add the conditions to play a round
Luckily there's only three comparisons that need
to be made between the two inputs (computer
, user). Because the the variables are strings
it means that the comparison has to be very
literal comparing strings against one another and
not using comparison operations but logic
operations instead. a message is logged to the
console for the inner and I made it so that
the message always displays the inputs
using template literals (this saves a lot of
space).