Feat: Change method for computerChoice function
In reviewing code to help with a solution towards looping the game
I found this more compact and sensible method of using an array
for holding the computer 'choices'. It's cutting down the amount
of lines of code, but it does the same exact thing as my previous
version of the function. Something I think I glossed over was that
the value 1 can never be a value of Math.random(), so my
<= 1 wasn't needed, it could just be < than 1. What this new
method has helped me understand is that array items can be selected
through their values within that array. Math.floor(Math.random() * 3)
is rounding down to the nearest whole number of the random value (0 to
 0.999etc.) then multiplying that by 3. Because floor is always rounding
 down, and that Math.random() will never be 1, it will always return
 0 when applying that function to it. By multiplying the result by 3
we are limiting the max number it can be. Potentially it could end up
returning 2.777 etc, but this is always rounded down by the Math.float,
which limits it to 0 to 2. This being the same index as the array.

computerChoice is then equal to a random index of choiceArray by
parsing in our value randomNum.

Arrays were a bit confusing before and one of the first things I
tried for this game, but I couldn't figure it out at the time how
to return a value from that array.