Fix: Fix game loop
Beforehand I had both the consts humanSelection & computerSelection
outside of the loop. I thought that having playRound() in the for
loop would work in that it plays through what already worked once,
but 5 times through this time when playGame() is called. The reason
that this didn't work off the bat was that the values of humanChoice &
computerChoice were fixed after the functions had been called once.
Even though the new consts were being added as arguements to
playRound() within the loop, their values were still the same
because the functions were being called outside of it once.

So computerSelection needed to be added to the loop before each
playRound() (so it went above it, as the program reads the loop top
to bottom and then loops again based on its conditons).
computerSelection now gets a new value assigned to it based on what
the function returns. humanSelection is the same, it gets a new
prompt value that it returns due to it equalling the return of that
function. Prompt is a modal window, which pauses the script to accept
user text input and this is needed in this case for one to stop the
script/scripts and also get a new choice from the user.