Feat: Update the conditions for the game
Instead of it playing 5 rounds overall, I wanted it to play the game
based on the upper limits of the score. Initially I used an ||
operator, this didn't work because it would go past the wanted limit.
Until both statements were false it could still progress, so either
the users or computers score could go beyond the limit by just
winning more rounds before the other got to the limit. By using &&
it would return the first falsey value, which would stop the program.
So when either score was less than 2 it would continue the loop,
but when the score hits 2 it stops looping and finsishes the game
in that way.